+++ lib/gamification/common/services/user_service.dart
@@ -42,10 +42,13 @@ class GamificationUserService with ChangeNotifier {
   /// Ride DAOs to access rides.
   RideSummaryDao get rideDao => AppDatabase.instance.rideSummaryDao;
 
+  /// The user profile for the gamificaiton functionality.
   UserProfile? get profile => _profile;
 
+  /// List of keys of the features enabled by the user.
   List<String> get enabledFeatures => _enabledFeatures;
 
+  /// List of keys of the features disabled by the user.
   List<String> get disabledFeatures =>
       gamificationFeatures.whereNot((feature) => enabledFeatures.contains(feature)).toList();
 
@@ -76,7 +79,7 @@ class GamificationUserService with ChangeNotifier {
     if (!(await _prefs?.setBool(gamificationEnabledKey, true) ?? false)) return false;
     // Start the database stream of rides, to update the profile data accordingly.
     startDatabaseStream();
-    sendProfileData();
+    sendProfileDataToBackend();
     return true;
   }
 
@@ -119,24 +122,27 @@ class GamificationUserService with ChangeNotifier {
   /// Returns true, if a given string key is inside of the list of selected game prefs.
   bool isFeatureEnabled(String key) => _enabledFeatures.contains(key);
 
+  /// Enabel the feature for a given key.
   Future<void> enableFeature(String key) async {
     if (_enabledFeatures.contains(key)) return;
     _enabledFeatures.add(key);
     _prefs ??= await SharedPreferences.getInstance();
     _prefs!.setStringList(enabledFeatureListKey, _enabledFeatures);
-    sendProfileData();
+    sendProfileDataToBackend();
     notifyListeners();
   }
 
+  /// Disable the feature with a given key.
   void disableFeature(String key) async {
     if (!_enabledFeatures.contains(key)) return;
     _enabledFeatures.remove(key);
     _prefs ??= await SharedPreferences.getInstance();
     _prefs!.setStringList(enabledFeatureListKey, _enabledFeatures);
-    sendProfileData();
+    sendProfileDataToBackend();
     notifyListeners();
   }
 
+  /// Move feature up in the feature-list, to change its position on the home view.
   void moveFeatureUp(String key) {
     if (_enabledFeatures.firstOrNull == key) return;
     int index = _enabledFeatures.indexOf(key);
@@ -145,6 +151,7 @@ class GamificationUserService with ChangeNotifier {
     notifyListeners();
   }
 
+  /// Move feature down in the feature-list, to change its position on the home view.
   void moveFeatureDown(String key) {
     if (_enabledFeatures.lastOrNull == key) return;
     int index = _enabledFeatures.indexOf(key);
@@ -153,6 +160,7 @@ class GamificationUserService with ChangeNotifier {
     notifyListeners();
   }
 
+  /// Reset the whole gamification profile of the user.
   Future<void> reset() async {
     var prefs = await SharedPreferences.getInstance();
     _profile = null;
@@ -160,16 +168,18 @@ class GamificationUserService with ChangeNotifier {
     prefs.remove(userProfileKey);
     prefs.remove(gamificationEnabledKey);
     prefs.remove(enabledFeatureListKey);
-    sendProfileData();
+    sendProfileDataToBackend();
     notifyListeners();
   }
 
-  Map<String, dynamic> get profilData => {
-        'gamificationEnabled': _profile != null,
-        'challengesEnabled': isFeatureEnabled(challengesFeatureKey),
-        'statisticsEnabled': isFeatureEnabled(statisticsFeatureKey),
-        'communityEnabled': false,
-      };
-
-  Future<void> sendProfileData() => getIt<EvaluationDataService>().sendJsonToAddress('settings/post/', profilData);
+  /// Send the users profile settings to the backend.
+  void sendProfileDataToBackend() {
+    var profilData = {
+      'gamificationEnabled': _profile != null,
+      'challengesEnabled': isFeatureEnabled(challengesFeatureKey),
+      'statisticsEnabled': isFeatureEnabled(statisticsFeatureKey),
+      'communityEnabled': false,
+    };
+    getIt<EvaluationDataService>().sendJsonToAddress('settings/post/', profilData);
+  }
 }
